+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+**Features**:
+
 - `normalize_performance_score` now handles `PerformanceScoreProfile` configs with zero weight components and component weight sums of any number greater than 0. ([#2756](https://github.com/getsentry/relay/pull/2756))
+- Return global config ready status to downstream relays  ([#2765](https://github.com/getsentry/relay/pull/2765))
 
 ## 23.11.1